@@ -4,6 +4,13 @@ import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
 import org.threeten.bp.LocalDate;
 import org.springframework.validation.annotation.Validated;
 import javax.validation.Valid;
@@ -15,168 +22,178 @@ import javax.validation.constraints.*;
 @Validated
 @javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-08-21T22:00:36.997790207Z[GMT]")
 
+@Entity
+@Table(name = "Users")
+public class User {
+	@Id
+	@Column(name = "ID")
+	@GeneratedValue(strategy = GenerationType.IDENTITY) // auto increment
+	@JsonProperty("id")
+	private Long id = null;
 
-public class User   {
-  @JsonProperty("id")
-  private Long id = null;
+	@Column(name = "NAME")
+	@JsonProperty("name")
+	private String name = null;
 
-  @JsonProperty("name")
-  private String name = null;
+	@Column(name = "BIRTH_DATE")
+	@JsonProperty("birthDate")
+	private LocalDate birthDate = null;
 
-  @JsonProperty("birthDate")
-  private LocalDate birthDate = null;
+	@Column(name = "EMAIL")
+	@JsonProperty("email")
+	private String email = null;
 
-  @JsonProperty("email")
-  private String email = null;
+	@Column(name = "STATUS")
+	@JsonProperty("status")
+	private String status = null;
 
-  @JsonProperty("status")
-  private String status = null;
+	public User id(Long id) {
+		this.id = id;
+		return this;
+	}
 
-  public User id(Long id) {
-    this.id = id;
-    return this;
-  }
+	/**
+	 * Get id
+	 * 
+	 * @return id
+	 **/
+	@Schema(example = "1", required = true, description = "")
+	@NotNull
 
-  /**
-   * Get id
-   * @return id
-   **/
-  @Schema(example = "1", required = true, description = "")
-      @NotNull
+	public Long getId() {
+		return id;
+	}
 
-    public Long getId() {
-    return id;
-  }
+	public void setId(Long id) {
+		this.id = id;
+	}
 
-  public void setId(Long id) {
-    this.id = id;
-  }
+	public User name(String name) {
+		this.name = name;
+		return this;
+	}
 
-  public User name(String name) {
-    this.name = name;
-    return this;
-  }
+	/**
+	 * Get name
+	 * 
+	 * @return name
+	 **/
+	@Schema(example = "Rodrigo Manrique da Penha", required = true, description = "")
+	@NotNull
 
-  /**
-   * Get name
-   * @return name
-   **/
-  @Schema(example = "Rodrigo Manrique da Penha", required = true, description = "")
-      @NotNull
+	public String getName() {
+		return name;
+	}
 
-    public String getName() {
-    return name;
-  }
+	public void setName(String name) {
+		this.name = name;
+	}
 
-  public void setName(String name) {
-    this.name = name;
-  }
+	public User birthDate(LocalDate birthDate) {
+		this.birthDate = birthDate;
+		return this;
+	}
 
-  public User birthDate(LocalDate birthDate) {
-    this.birthDate = birthDate;
-    return this;
-  }
+	/**
+	 * Get birthDate
+	 * 
+	 * @return birthDate
+	 **/
+	@Schema(example = "Tue Jul 04 00:00:00 GMT 2000", required = true, description = "")
+	@NotNull
 
-  /**
-   * Get birthDate
-   * @return birthDate
-   **/
-  @Schema(example = "Tue Jul 04 00:00:00 GMT 2000", required = true, description = "")
-      @NotNull
+	@Valid
+	public LocalDate getBirthDate() {
+		return birthDate;
+	}
 
-    @Valid
-    public LocalDate getBirthDate() {
-    return birthDate;
-  }
+	public void setBirthDate(LocalDate birthDate) {
+		this.birthDate = birthDate;
+	}
 
-  public void setBirthDate(LocalDate birthDate) {
-    this.birthDate = birthDate;
-  }
+	public User email(String email) {
+		this.email = email;
+		return this;
+	}
 
-  public User email(String email) {
-    this.email = email;
-    return this;
-  }
+	/**
+	 * Get email
+	 * 
+	 * @return email
+	 **/
+	@Schema(example = "rodrigomanriquep@gmail.com", required = true, description = "")
+	@NotNull
 
-  /**
-   * Get email
-   * @return email
-   **/
-  @Schema(example = "rodrigomanriquep@gmail.com", required = true, description = "")
-      @NotNull
+	public String getEmail() {
+		return email;
+	}
 
-    public String getEmail() {
-    return email;
-  }
+	public void setEmail(String email) {
+		this.email = email;
+	}
 
-  public void setEmail(String email) {
-    this.email = email;
-  }
+	public User status(String status) {
+		this.status = status;
+		return this;
+	}
 
-  public User status(String status) {
-    this.status = status;
-    return this;
-  }
+	/**
+	 * Get status
+	 * 
+	 * @return status
+	 **/
+	@Schema(example = "ACTIVE", required = true, description = "")
+	@NotNull
 
-  /**
-   * Get status
-   * @return status
-   **/
-  @Schema(example = "ACTIVE", required = true, description = "")
-      @NotNull
+	public String getStatus() {
+		return status;
+	}
 
-    public String getStatus() {
-    return status;
-  }
+	public void setStatus(String status) {
+		this.status = status;
+	}
 
-  public void setStatus(String status) {
-    this.status = status;
-  }
+	@Override
+	public boolean equals(java.lang.Object o) {
+		if (this == o) {
+			return true;
+		}
+		if (o == null || getClass() != o.getClass()) {
+			return false;
+		}
+		User user = (User) o;
+		return Objects.equals(this.id, user.id) && Objects.equals(this.name, user.name)
+				&& Objects.equals(this.birthDate, user.birthDate) && Objects.equals(this.email, user.email)
+				&& Objects.equals(this.status, user.status);
+	}
 
+	@Override
+	public int hashCode() {
+		return Objects.hash(id, name, birthDate, email, status);
+	}
 
-  @Override
-  public boolean equals(java.lang.Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    User user = (User) o;
-    return Objects.equals(this.id, user.id) &&
-        Objects.equals(this.name, user.name) &&
-        Objects.equals(this.birthDate, user.birthDate) &&
-        Objects.equals(this.email, user.email) &&
-        Objects.equals(this.status, user.status);
-  }
+	@Override
+	public String toString() {
+		StringBuilder sb = new StringBuilder();
+		sb.append("class User {\n");
 
-  @Override
-  public int hashCode() {
-    return Objects.hash(id, name, birthDate, email, status);
-  }
+		sb.append("    id: ").append(toIndentedString(id)).append("\n");
+		sb.append("    name: ").append(toIndentedString(name)).append("\n");
+		sb.append("    birthDate: ").append(toIndentedString(birthDate)).append("\n");
+		sb.append("    email: ").append(toIndentedString(email)).append("\n");
+		sb.append("    status: ").append(toIndentedString(status)).append("\n");
+		sb.append("}");
+		return sb.toString();
+	}
 
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class User {\n");
-    
-    sb.append("    id: ").append(toIndentedString(id)).append("\n");
-    sb.append("    name: ").append(toIndentedString(name)).append("\n");
-    sb.append("    birthDate: ").append(toIndentedString(birthDate)).append("\n");
-    sb.append("    email: ").append(toIndentedString(email)).append("\n");
-    sb.append("    status: ").append(toIndentedString(status)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(java.lang.Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
+	/**
+	 * Convert the given object to string with each line indented by 4 spaces
+	 * (except the first line).
+	 */
+	private String toIndentedString(java.lang.Object o) {
+		if (o == null) {
+			return "null";
+		}
+		return o.toString().replace("\n", "\n    ");
+	}
 }
